Update 03-primer_stats.Rmd
First, thank you for this wonderful book and for making all of your code accessible. 

I am struggling with the wording of Chapter 3, Line 189. It seems that since the upper and lower bounds of the difference in sample means do not include zero, we cannot have even 5% confidence that the difference in population means is zero. I suspect this is my misunderstanding rather than an error in the text, but I cannot square the circle in my head, so I am raising it here.